requirements.txt
sqlalchemy 
psycopg2 
pandas
python-dotenv
pandas-profiling
fastapi
uvicorn
pydantic==1.10.9

